[ngl/render/aeps/platform] ngl project/unproject/device-matrix/dump/callback/frame-count/depth-tex/clip-result typed, shader reg unsigned int*, nglIsSphereVisible frustum form, defaultFileName QBDB, InitDefaultPak apkFile*, FEManager object; LNK 4599->4507

+++ src/render/cdDecalShader.h
@@ -65,8 +65,8 @@ namespace cdDecalFullbrightPixel {
 // ============================================================================
 // Externs
 // ============================================================================
-extern void nglDxRegisterVShader(unsigned long* VS, const unsigned int* Microcode);  // ngl_dx_shader.o
-extern void nglDxRegisterPShader(unsigned long** PS, const unsigned int* Microcode);  // ngl_dx_shader.o
+extern void nglDxRegisterVShader(unsigned int* VS, const unsigned int* Microcode);  // ngl_dx_shader.o
+extern void nglDxRegisterPShader(unsigned int** PS, const unsigned int* Microcode);  // ngl_dx_shader.o
 
 extern cdDecalShader* gCDDecalShader;   // @0x14CDA4C
 extern cdDecalShader* g_cdDecalShader;  // @0x14CDA48
