+++ src/render/cdWorldVertexLitShader.h
@@ -66,8 +66,8 @@ namespace cdWorldVertexLitFullbrightPixel {
 // ============================================================================
 // Externs
 // ============================================================================
-extern void nglDxRegisterVShader(unsigned long* VS, const unsigned int* Microcode);  // ngl_dx_shader.o
-extern void nglDxRegisterPShader(unsigned long** PS, const unsigned int* Microcode);  // ngl_dx_shader.o
+extern void nglDxRegisterVShader(unsigned int* VS, const unsigned int* Microcode);  // ngl_dx_shader.o
+extern void nglDxRegisterPShader(unsigned int** PS, const unsigned int* Microcode);  // ngl_dx_shader.o
 
 extern cdWorldVertexLitShader* gCDWorldVertexLitShader;    // @0x10DE5DC
 extern cdWorldVertexLitShader* g_cdWorldVertexLitShader;   // @0x10DE5D8
